Farm (#1)
* Starting to implement the CRDT

We probably want to replace the map with a LRU as it provides a lot more
useful information for the CRDT.

* Implementing a LRU cache for the local node

This makes it much easier to enable cache evication for old cache
items.

* Better communication between the cache and the store

* Add a Value to the store

This make this more a KVS rather than just a coherence engine

* Adding nodes remove + virtual tests

Adding store mock file, so that we can better test the virtual node

* Adding nodeset tests

* Adding the circuit breaker to the farm

* Fixing the union and difference function

We now only show union sets with consensus members. This is a breaking
change, where previously we would always show all members in the union,
even though they would have only be replicated to potentially one node.

This now fixes that and only shows items that have made consensus.